requirements.txt
alembic==1.8.1
altair==4.2.0
asgiref==3.5.2
attrs==22.1.0
azure-ai-formrecognizer==3.2.0
azure-common==1.1.28
azure-core==1.26.1
backoff==1.11.1
backports.zoneinfo==0.2.1
blinker==1.5
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
cloudpickle==2.2.0
commonmark==0.9.1
construct==2.5.3
cycler==0.11.0
databricks-cli==0.17.3
decorator==5.1.1
dill==0.3.6
Django==4.1.3
django-haystack==3.2.1
docker==6.0.1
docopt==0.6.2
elasticsearch==7.17.7
entrypoints==0.4
farm-haystack==1.10.0
filelock==3.8.0
Flask==2.2.2
fonttools==4.33.3
future==0.18.2
gitdb==4.0.9
GitPython==3.1.29
greenlet==2.0.1
gunicorn==20.1.0
haystack==0.42
huggingface-hub==0.11.0
idna==3.4
importlib-metadata==5.0.0
importlib-resources==5.10.0
inflect==6.0.2
isodate==0.6.1
itsdangerous==2.1.2
jarowinkler==1.2.3
Jinja2==3.1.2
joblib==1.1.0
jsonschema==4.17.0
kiwisolver==1.4.2
langdetect==1.0.9
llvmlite==0.39.1
lxml==4.9.1
Mako==1.2.4
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlip==0.2
mlflow==2.0.1
mmh3==3.0.0
monotonic==1.6
more-itertools==9.0.0
msrest==0.7.1
networkx==2.8.8
nltk==3.7
num2words==0.5.12
numba==0.56.4
numpy==1.22.3
oauthlib==3.2.2
packaging==21.3
pandas==1.4.2
pefile==2022.5.30
Pillow==9.1.1
pkgutil_resolve_name==1.3.10
posthog==2.2.0
protobuf==3.20.3
pyarrow==10.0.0
pydantic==1.10.2
pydeck==0.8.0
Pygments==2.13.0
PyJWT==2.6.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.2
python-dateutil==2.8.2
python-docx==0.8.11
python-ptrace==0.9.8
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
quantulum3==0.7.11
querystring-parser==1.2.4
rapidfuzz==2.7.0
regex==2022.10.31
requests==2.28.1
requests-oauthlib==1.3.1
rich==12.6.0
scikit-learn==1.1.1
scipy==1.8.1
seaborn==0.11.2
semver==2.13.0
sentence-transformers==2.2.2
sentencepiece==0.1.97
seqeval==1.2.2
shap==0.41.0
six==1.16.0
sklearn==0.0
slicer==0.0.7
smmap==5.0.0
SQLAlchemy==1.4.44
sqlparse==0.4.3
streamlit==1.15.0
tabulate==0.9.0
threadpoolctl==3.1.0
tika==1.24
tokenizers==0.12.1
toml==0.10.2
toolz==0.12.0
torch==1.12.1
torchvision==0.13.1
tornado==6.2
tqdm==4.64.1
transformers==4.21.2
typing_extensions==4.4.0
tzdata==2022.6
tzlocal==4.2
urllib3==1.26.12
validators==0.20.0
watchdog==2.1.9
websocket-client==1.4.2
Werkzeug==2.2.2
zipp==3.10.0
